Refactor to use controller classes which means one thing can be locked instead of many
